Fix the error in the "verify-clean-git-stage" rule of the Makefile
The first mistake was that `make` ate the `$`, have to use `$$` to
escape.

The second mistake with the old version, which was in the form of:

```bash
[ -n "$(git status --porcelain)" ] || \
{ echo "ERROR"; git status --porcelain; exit 1; }
```

was that if `[ -n "$(git status --porcelain)" ]` is `false` (which
would mean the staging area is clean and everything is good to move
on), it behaves like a failed process and exits (that's what `false`
is and does for such statement in bash), hence failing the the whole
bash process running the current line, and consequently the make
process.

The current version lets the `if` catch the potential fail (i.e. the
state being `false`). Alternatively, instead of using `if` to catch
the `false`, I could let it lead to an idle process with an "or" by
the use of `||:` where `:` does nothing. Something like:

```bash
[ -n "$$(git status --porcelain)" ] && \
{ echo "ERROR"; git status --porcelain; exit 1; } \
|| :
```