Huge refactoring
With the former structure of the repo, there was plenty of name
clashes and issues with pylint.

This commit established a uniform structure for the repo. Accordingly
each individual [formerly independent and stand alone] sandboxes is
turned into a separate module under `libs` and each gets its own
"entry" script under `entry_points`.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>